Queryを別ファイルへ記述(Articlesページ)
