+++ src/pages/PageQuery/articlesQuery.ts
@@ -1,0 +1,20 @@
+import { graphql, useStaticQuery } from 'gatsby'
+import type { ArticlesPageQuery } from 'types/graphql-types'
+
+export const UseArticlesPageQuery = () => {
+  const data = useStaticQuery<ArticlesPageQuery>(graphql`
+    query ArticlesPage {
+      contentfulArticles {
+        title
+        createdArticleDateJP: createdArticleDate(formatString: "YYYY年MM月DD日")
+        createdArticleDate
+        category {
+          category
+          categorySlug
+          id
+        }
+      }
+    }
+  `)
+  return data
+}
